requirements.txt
# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
neo4j>=5.0.0

# Data Models & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Graph Analysis
networkx>=3.0

# Report Generation
Jinja2>=3.1.0

# Environment Variables
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# CLI
click>=8.1.0
rich>=13.0.0
